+++ frontend/src/themes/header/header.jsx
@@ -1,10 +1,9 @@
-import React, { useState } from "react"
+import React from "react"
 import { makeStyles } from "@material-ui/core/styles"
-import { Button, Typography, Container, Link } from "@material-ui/core"
-import { useTranslation } from "react-i18next"
+import { Typography, Container, Link } from "@material-ui/core"
 import logoCom2kube from "../../assets/images/logo_transparent.png"
-import i18n from "../../i18n"
 import { Logo, Name } from "./style"
+import Links from "./links/links"
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -23,9 +22,6 @@ const useStyles = makeStyles((theme) => ({
     alignItems: "center",
     alignContent: "center"
   },
-  button: {
-    margin: theme.spacing(1)
-  },
   navLinks: {
     display: "flex",
     justifyContent: "flex-end",
@@ -35,19 +31,7 @@ const useStyles = makeStyles((theme) => ({
 }))
 
 function Header() {
-  const [isHidden, isHiddenState] = useState(false)
   const classes = useStyles()
-
-  const changeLanguage = (lng) => {
-    i18n.changeLanguage(lng)
-  }
-
-  const langState = (lang) => {
-    changeLanguage(lang)
-    isHiddenState(!isHidden)
-  }
-  const { t } = useTranslation()
-
   return (
     <div className={classes.root}>
       <Container className={classes.nameLogo}>
@@ -69,40 +53,7 @@ function Header() {
           </Typography>
         </Name>
       </Container>
-      <Container className={classes.navLinks}>
-        <Button color="inherit" href="/about">
-          {t("about.menu")}
-        </Button>
-        <Button
-          color="inherit"
-          onClick={() => window.open("https://github.com/CB-GJ/Com2Kube", "_blank")}
-        >
-          Github
-        </Button>
-        {isHidden ? (
-          <Button
-            onClick={() => {
-              langState("en")
-            }}
-            type="button"
-            alt="english lang"
-            className={classes.button}
-          >
-            English
-          </Button>
-        ) : (
-          <Button
-            onClick={() => {
-              langState("fr")
-            }}
-            type="button"
-            alt="french lang"
-            className={classes.button}
-          >
-            French
-          </Button>
-        )}
-      </Container>
+      <Links />
     </div>
   )
 }
